models/requirements.txt
aiofiles==24.1.0
aiohttp==3.9.5
aiolimiter==1.2.1
aiosignal==1.3.2
aiosqlite==0.20.0
alembic==1.14.0
annotated-types==0.7.0
anthropic==0.42.0
anyio==4.7.0
appdirs==1.4.4
asgiref==3.8.1
assemblyai==0.36.0
astra-assistants==2.2.7
astrapy==1.5.2
asttokens==3.0.0
asyncer==0.0.8
atlassian-python-api==3.41.16
attrs==24.3.0
auth0-python==4.7.2
Authlib==1.3.1
backoff==2.2.1
bce-python-sdk==0.9.25
bcrypt==4.0.1
beautifulsoup4==4.12.3
blinker==1.9.0
boto3==1.34.162
botocore==1.34.162
build==1.2.2.post1
cachetools==5.5.0
cassandra-driver==3.29.2
cassio==0.1.10
cattrs==24.1.2
certifi==2024.12.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
chroma-hnswlib==0.7.6
chromadb==0.5.23
clevercsv==0.8.3
click==8.1.8
clickhouse-connect==0.7.19
cohere==5.13.4
coloredlogs==15.0.1
colorlog==6.9.0
composio_core==0.5.42
composio_langchain==0.5.42
coolname==2.2.0
crewai==0.80.0
crewai-tools==0.17.0
cryptography==43.0.3
dataclasses-json==0.6.7
datasets==2.14.4
decorator==5.1.1
deepdiff==8.1.1
defusedxml==0.7.1
Deprecated==1.2.15
deprecation==2.1.0
dill==0.3.7
diskcache==5.6.3
distro==1.9.0
dnspython==2.7.0
docker==7.1.0
docstring_parser==0.16
docx2txt==0.8
dspy==2.5.7
dspy-ai==2.5.43
duckdb==1.1.3
duckduckgo_search==6.4.2
durationpy==0.9
e2b==1.0.5
e2b-code-interpreter==1.0.3
ecdsa==0.19.0
elastic-transport==8.15.1
elasticsearch==8.17.0
embedchain==0.1.126
emoji==2.14.0
environs==9.5.0
Events==0.5
executing==2.1.0
faiss-cpu==1.9.0.post1
fake-useragent==1.5.1
fastapi==0.115.6
fastapi-pagination==0.12.34
fastavro==1.10.0
filelock==3.16.1
filetype==1.2.0
firecrawl-py==1.6.8
FLAML==2.3.3
Flask==3.1.0
Flask-Cors==5.0.0
flatbuffers==24.12.23
frozendict==2.4.6
frozenlist==1.5.0
fsspec==2024.12.0
future==1.0.0
geomet==0.2.1.post1
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.10
google-api-core==2.24.0
google-api-python-client==2.159.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-cloud-aiplatform==1.75.0
google-cloud-bigquery==3.27.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.14.0
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-generativeai==0.8.3
google-resumable-media==2.7.2
google_search_results==2.4.2
googleapis-common-protos==1.66.0
googlesearch-python==1.3.0
gotrue==2.11.1
gptcache==0.1.44
grandalf==0.8
greenlet==3.1.1
groq==0.13.1
grpc-google-iam-v1==0.13.1
grpcio==1.68.1
grpcio-health-checking==1.68.1
grpcio-status==1.68.1
grpcio-tools==1.68.1
gunicorn==23.0.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
html5lib==1.1
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx==0.27.0
httpx-sse==0.4.0
huggingface-hub==0.27.0
humanfriendly==10.0
hyperframe==6.0.1
idna==3.10
ijson==3.3.0
importlib_metadata==8.5.0
importlib_resources==6.4.5
inflection==0.5.1
iniconfig==2.0.0
instructor==1.3.4
ipython==8.31.0
itsdangerous==2.2.0
jaraco.context==6.0.1
jedi==0.19.2
Jinja2==3.1.5
jiter==0.4.2
jmespath==1.0.1
joblib==1.4.2
jq==1.8.0
json_repair==0.35.0
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpickle==4.0.1
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kubernetes==31.0.0
lancedb==0.17.0
langchain==0.3.13
langchain-anthropic==0.3.1
langchain-astradb==0.5.2
langchain-aws==0.2.7
langchain-chroma==0.1.4
langchain-cohere==0.3.4
langchain-community==0.3.13
langchain-core==0.3.28
langchain-elasticsearch==0.3.0
langchain-experimental==0.3.4
langchain-google-calendar-tools==0.0.1
langchain-google-community==2.0.3
langchain-google-genai==2.0.7
langchain-google-vertexai==2.0.7
langchain-groq==0.2.2
langchain-milvus==0.1.7
langchain-mistralai==0.2.4
langchain-mongodb==0.3.0
langchain-nvidia-ai-endpoints==0.3.7
langchain-ollama==0.2.1
langchain-openai==0.2.14
langchain-pinecone==0.2.0
langchain-text-splitters==0.3.4
langchain-unstructured==0.1.5
langchainhub==0.1.21
langflow==1.1.1
langflow-base==0.1.1
langfuse==2.53.9
langsmith==0.1.147
langwatch==0.1.16
lark==1.2.2
litellm==1.56.4
loguru==0.7.3
lsprotocol==2023.0.1
lxml==5.3.0
lz4==4.3.3
magicattr==0.1.6
Mako==1.3.8
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mem0ai==0.1.38
metal_sdk==2.5.1
metaphor-python==0.1.23
milvus-lite==2.4.11
mmh3==5.0.1
monotonic==1.6
more-itertools==10.5.0
motor==3.5.1
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.15
multitasking==0.0.11
mypy-extensions==1.0.0
nanoid==2.0.0
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
nodeenv==1.9.1
numexpr==2.10.2
numpy==1.26.4
oauthlib==3.2.2
ollama==0.3.3
onnxruntime==1.20.1
openai==1.58.1
opensearch-py==2.8.0
opentelemetry-api==1.29.0
opentelemetry-exporter-otlp-proto-common==1.29.0
opentelemetry-exporter-otlp-proto-grpc==1.29.0
opentelemetry-exporter-otlp-proto-http==1.29.0
opentelemetry-exporter-prometheus==0.50b0
opentelemetry-instrumentation==0.50b0
opentelemetry-instrumentation-asgi==0.50b0
opentelemetry-instrumentation-fastapi==0.50b0
opentelemetry-proto==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-semantic-conventions==0.50b0
opentelemetry-util-http==0.50b0
optuna==4.1.0
orderly-set==5.2.3
orjson==3.10.0
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.2
pandas==2.2.2
parameterized==0.9.0
paramiko==3.5.0
parso==0.8.4
passlib==1.7.4
peewee==3.17.8
pexpect==4.9.0
pgvector==0.3.6
pillow==10.4.0
pinecone-client==5.0.1
pinecone-plugin-inference==1.1.0
pinecone-plugin-interface==0.0.7
platformdirs==4.3.6
pluggy==1.5.0
portalocker==2.10.1
postgrest==0.16.11
posthog==3.7.4
primp==0.9.2
prometheus_client==0.21.1
prompt_toolkit==3.0.48
propcache==0.2.1
proto-plus==1.25.0
protobuf==5.29.3
psycopg==3.2.3
psycopg-binary==3.2.3
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
py==1.11.0
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyautogen==0.5.3
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.7.4
pydantic-settings==2.4.0
pydantic_core==2.18.4
Pygments==2.18.0
PyJWT==2.10.1
pylance==0.20.0
pymilvus==2.4.9
pymongo==4.10.1
PyNaCl==1.5.0
pyparsing==3.2.1
pypdf==5.1.0
pyperclip==1.9.0
PyPika==0.48.9
pyproject_hooks==1.2.0
pyright==1.1.391
pysbd==0.3.4
Pysher==1.0.8
PySocks==1.7.1
pytest==8.3.4
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-jose==3.3.0
python-lsp-jsonrpc==1.1.2
python-multipart==0.0.20
pytube==15.0.0
pytz==2024.2
pyvis==0.3.2
PyYAML==6.0.2
qdrant-client==1.9.2
qianfan==0.3.5
ragstack-ai-knowledge-store==0.2.1
realtime==1.0.6
redis==5.2.1
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
retry==0.9.2
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
ruff==0.8.4
s3transfer==0.10.4
schema==0.7.7
selenium==4.27.1
semver==3.0.2
sentry-sdk==2.19.2
setuptools==75.6.0
shapely==2.0.6
shellingham==1.5.4
simsimd==6.2.1
six==1.17.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
spider-client==0.1.25
SQLAlchemy==2.0.36
sqlmodel==0.0.18
stack-data==0.6.3
starlette==0.41.3
storage3==0.7.7
StrEnum==0.4.15
structlog==24.4.0
supabase==2.6.0
supafunc==0.5.1
sympy==1.13.3
tabulate==0.9.0
tenacity==8.5.0
termcolor==2.5.0
tiktoken==0.7.0
tokenizers==0.20.3
toml==0.10.2
tomli==2.2.1
tomli_w==1.1.0
tqdm==4.67.1
traitlets==5.14.3
tree-sitter==0.23.2
tree-sitter-python==0.23.6
trio==0.28.0
trio-websocket==0.11.1
typer==0.15.1
types-cachetools==5.5.0.20240820
types-requests==2.32.0.20241016
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
ujson==5.10.0
uncurl==0.0.11
unstructured-client==0.25.9
upstash-vector==0.6.0
uritemplate==4.1.1
urllib3==2.3.0
uuid6==2024.7.10
uv==0.5.13
uvicorn==0.34.0
uvloop==0.21.0
validators==0.34.0
watchfiles==1.0.3
wcwidth==0.2.13
weaviate-client==4.9.6
webencodings==0.5.1
websocket-client==1.8.0
websockets==12.0
Werkzeug==3.1.3
wikipedia==1.4.0
wolframalpha==5.1.3
wrapt==1.17.0
wsproto==1.2.0
xmltodict==0.14.2
xxhash==3.5.0
yarl==1.18.3
yfinance==0.2.51
youtube-transcript-api==0.6.3
zep-python==2.0.2
zipp==3.21.0
zstandard==0.23.0
gunicorn
